Added new configuration for grammar generation.
Features a new word-level grammar, which does not feature POS or
semantic symbols. The only knowledge captured in this grammar is word
boundaries and length.

This commit also includes a change in he output of grammars. They are no
longer output in a subfolder named with the password set it.