[fix] tree cut wasn't aware of sampling. cut was always calculated with entire sample
From email written to Julie:
My code was reading a cached version of the weighted (w/ occurrence counts) WordNet tree. This is because weighing the tree is costly, as it requires a full read of the db. But I was keeping a single cache for all grammars of a password set (e.g., RockYou). This means a grammar trained with a small sample was benefiting from tree cuts made with more information (because that one cache was built with the full database). I overlooked this when refactoring the code to deal with multiple grammars (it used to be monolithic).